requirements.txt
alembic==1.13.1
Authlib==1.1.0
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7
cloudinary==1.44.1
colorama==0.4.6
commonmark==0.9.1
cryptography==45.0.6
Deprecated==1.2.18
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
Flask==3.0.3
Flask-Admin==1.6.1
Flask-Caching==2.3.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
greenlet==3.2.4
gunicorn==22.0.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
limits==5.5.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
pillow==10.4.0
pipreqs==0.4.13
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.31.0
rich==13.9.4
setuptools==80.9.0
six==1.17.0
SQLAlchemy==1.4.52
stripe==12.4.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.0.0
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
yarg==0.1.10
Flask-Bcrypt==1.0.1
Flask-SocketIO
python-socketio
python-engineio
simple-websocket
wsproto
h11
bidict
flask-mongoengine-3>=1.1.0
mongoengine==0.29.1
eventlet
requests-oauthlib==2.0.0
oauthlib==3.2.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.133.0
flask-apscheduler

